Move needed types out of spirv only
