+++ naga/src/back/spv/ray.rs
@@ -5,10 +5,10 @@ Generating SPIR-V for ray query operations.
 use alloc::{vec, vec::Vec};
 
 use super::{
-    Block, BlockContext, Function, FunctionArgument, Instruction, LookupFunctionType, NumericType,
-    Writer,
+    Block, BlockContext, Function, FunctionArgument, Instruction, LookupFunctionType,
+    LookupRayQueryFunction, NumericType, Writer,
 };
-use crate::{arena::Handle, back::spv::LookupRayQueryFunction};
+use crate::{arena::Handle, back::RayQueryPoint};
 
 /// helper function to check if a particular flag is set in a u32.
 fn write_ray_flags_contains_flags(
@@ -190,13 +190,13 @@ impl Writer {
             self,
             &mut block,
             loaded_ray_query_tracker_id,
-            super::RayQueryPoint::PROCEED.bits(),
+            RayQueryPoint::PROCEED.bits(),
         );
         let finished_proceed_id = write_ray_flags_contains_flags(
             self,
             &mut block,
             loaded_ray_query_tracker_id,
-            super::RayQueryPoint::FINISHED_TRAVERSAL.bits(),
+            RayQueryPoint::FINISHED_TRAVERSAL.bits(),
         );
         let proceed_finished_correct_id = if is_committed {
             finished_proceed_id
@@ -998,9 +998,8 @@ impl Writer {
             tmax_id,
         ));
 
-        let const_initialized = self.get_constant_scalar(crate::Literal::U32(
-            super::RayQueryPoint::INITIALIZED.bits(),
-        ));
+        let const_initialized =
+            self.get_constant_scalar(crate::Literal::U32(RayQueryPoint::INITIALIZED.bits()));
         valid_block
             .body
             .push(Instruction::store(init_tracker_id, const_initialized, None));
@@ -1089,7 +1088,7 @@ impl Writer {
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::INITIALIZED.bits(),
+                RayQueryPoint::INITIALIZED.bits(),
             );
 
             block.body.push(Instruction::selection_merge(
@@ -1116,10 +1115,10 @@ impl Writer {
             .push(Instruction::store(proceeded_id, has_proceeded, None));
 
         let add_flag_finished = self.get_constant_scalar(crate::Literal::U32(
-            (super::RayQueryPoint::PROCEED | super::RayQueryPoint::FINISHED_TRAVERSAL).bits(),
+            (RayQueryPoint::PROCEED | RayQueryPoint::FINISHED_TRAVERSAL).bits(),
         ));
         let add_flag_continuing =
-            self.get_constant_scalar(crate::Literal::U32(super::RayQueryPoint::PROCEED.bits()));
+            self.get_constant_scalar(crate::Literal::U32(RayQueryPoint::PROCEED.bits()));
 
         let add_flags_id = self.id_gen.next();
         valid_block.body.push(Instruction::select(
@@ -1226,13 +1225,13 @@ impl Writer {
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::PROCEED.bits(),
+                RayQueryPoint::PROCEED.bits(),
             );
             let finished_proceed_id = write_ray_flags_contains_flags(
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::FINISHED_TRAVERSAL.bits(),
+                RayQueryPoint::FINISHED_TRAVERSAL.bits(),
             );
 
             // Can't find anything to suggest double calling this function is invalid.
@@ -1501,13 +1500,13 @@ impl Writer {
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::PROCEED.bits(),
+                RayQueryPoint::PROCEED.bits(),
             );
             let finished_proceed_id = write_ray_flags_contains_flags(
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::FINISHED_TRAVERSAL.bits(),
+                RayQueryPoint::FINISHED_TRAVERSAL.bits(),
             );
             // Although it seems strange to call this twice, I (Vecvec) can't find anything to suggest double calling this function is invalid.
             let not_finished_id = self.id_gen.next();
@@ -1673,13 +1672,13 @@ impl Writer {
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::PROCEED.bits(),
+                RayQueryPoint::PROCEED.bits(),
             );
             let finished_proceed_id = write_ray_flags_contains_flags(
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::FINISHED_TRAVERSAL.bits(),
+                RayQueryPoint::FINISHED_TRAVERSAL.bits(),
             );
 
             let correct_finish_id = if is_committed {
@@ -1825,14 +1824,14 @@ impl Writer {
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::PROCEED.bits(),
+                RayQueryPoint::PROCEED.bits(),
             );
 
             let finished_proceed_id = write_ray_flags_contains_flags(
                 self,
                 &mut block,
                 initialized_tracker_id,
-                super::RayQueryPoint::FINISHED_TRAVERSAL.bits(),
+                RayQueryPoint::FINISHED_TRAVERSAL.bits(),
             );
 
             let not_finished_id = self.id_gen.next();
